fix: Apply image format when saving file in ImagickDriver
Certain functions like `resizeCanvas`, `overlay` actually re-assign
the `image` property, thus "losing" the format if it was set prior to
calling those transformations.

This can be avoided by ensuring `setFormat()` on the Imagick image(s)
is only called prior to saving the image, at which point the format
will always be correct when saving the file.

To do this, the format is stored as a property on the driver object.
(The GD driver already does it the same way, so I just made a few
adjustments to bring this driver in line with the GD driver.)