Improve DX by specifying namespace for GD functions
Why: some of these functions are optional and can be not available in the installed PHP.
When PHP couldn't find functions in the current and root namespaces, it reports about missing function in the current namespace.
As result, Error message is misleading.
Example: "Call to undefined function Spatie\Image\Drivers\Gd\imageavif()"